JS: add null guard to 'kill' method for ExternalProgramServer

+++ src/main/java/com/jaamsim/BasicObjects/ExternalProgramServer.java
@@ -68,7 +68,9 @@ public class ExternalProgramServer extends AbstractExternalProgram {
 	public void kill() {
 		super.kill();
 
-		killProcess();
+		if (process != null) {
+			killProcess();
+		}
 	}
 
 	private void startProcess() {
